fix: macOS type compatibility for libc fields
Fix build failures on macOS due to type differences in libc:
- msg_controllen is usize on Linux but u32 on macOS
- cmsg_len is usize on Linux but u32 on macOS
- TIOCSCTTY requires cast to c_ulong for ioctl on macOS

Added explicit casts that work correctly on both platforms.

+++ crates/cterm-core/src/native_pty.rs
@@ -223,7 +223,8 @@ impl NativePty {
         }
 
         // Set the slave as the controlling terminal
-        if libc::ioctl(slave_fd, libc::TIOCSCTTY, 0) < 0 {
+        // Cast needed: TIOCSCTTY type differs between Linux and macOS
+        if libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0) < 0 {
             libc::_exit(1);
         }
